Evaluation Report

+++ PayItForward/PayItForward/Pages/AdminProfile.aspx.cs
@@ -212,8 +212,8 @@ namespace PayItForward.Pages
             bool flag = false;
 
             float weight = 0.0f;
-            float low    = 0.0f;
-            float high   = 0.0f;
+            decimal low    = 0;
+            decimal high   = 0;
 
             if (itemName.Text   == null || itemName.Text.Trim().Length <= 0    || !(new Regex("^[A-Za-z0-9()' ]+$").IsMatch(itemName.Text.Trim())))
             {
@@ -223,11 +223,11 @@ namespace PayItForward.Pages
             {
                 flag = true;
             }
-            if (itemLow.Text    == null || itemLow.Text.Trim().Length <= 0     || !float.TryParse(itemLow.Text.Trim(), out low))
+            if (itemLow.Text    == null || itemLow.Text.Trim().Length <= 0     || !decimal.TryParse(itemLow.Text.Trim(), out low))
             {
                 flag = true;
             }
-            if (itemHigh.Text   == null || itemHigh.Text.Trim().Length <= 0    || !float.TryParse(itemHigh.Text.Trim(), out high))
+            if (itemHigh.Text   == null || itemHigh.Text.Trim().Length <= 0    || !decimal.TryParse(itemHigh.Text.Trim(), out high))
             {
                 flag = true;
             }
